+++ benchmarks/cuda/patterns/runge_kutta4.hpp
@@ -93,8 +93,12 @@ struct RungeKutta4Pattern
         state.ResumeTiming();
 
         rk4_kernel<<<
-            p.agents / p.block_size + (p.agents % p.block_size != 0 ? 1 : 0),
-            p.block_size>>>(device_agents, p.agents, p.steps, f);
+            static_cast<unsigned int>(
+                p.agents / p.block_size + (p.agents % p.block_size != 0 ? 1 : 0)
+            ),
+            static_cast<unsigned int>(p.block_size)>>>(
+            device_agents, p.agents, p.steps, f
+        );
 
         cudaErrorCheck(cudaGetLastError());
         cudaErrorCheck(cudaDeviceSynchronize());
